+++ numpy/_build_utils/common.py
